Move git helpers to dedicated file
