+++ packages/create-cloudflare/src/__tests__/common.test.ts
@@ -1,18 +1,12 @@
 import { spinner } from "@cloudflare/cli/interactive";
-import * as command from "helpers/command";
 import { SemVer } from "semver";
 import { getGlobalDispatcher, MockAgent, setGlobalDispatcher } from "undici";
 import { afterEach, beforeEach, describe, expect, test, vi } from "vitest";
 import { version as currentVersion } from "../../package.json";
-import { isGitConfigured } from "../common";
 import { isUpdateAvailable } from "../helpers/cli";
 
 vi.mock("process");
 vi.mock("@cloudflare/cli/interactive");
-
-function promisify<T>(value: T) {
-	return new Promise<T>((res) => res(value));
-}
 
 beforeEach(() => {
 	// we mock `spinner` to remove noisy logs from the test runs
@@ -21,40 +15,6 @@ beforeEach(() => {
 		update() {},
 		stop() {},
 	}));
-});
-
-describe("isGitConfigured", () => {
-	test("fully configured", async () => {
-		const spy = vi.spyOn(command, "runCommand");
-		spy.mockImplementation((cmd) =>
-			promisify(cmd.includes("email") ? "test@user.com" : "test user")
-		);
-		expect(await isGitConfigured()).toBe(true);
-	});
-
-	test("no name", async () => {
-		const spy = vi.spyOn(command, "runCommand");
-		spy.mockImplementation((cmd) =>
-			promisify(cmd.includes("email") ? "test@user.com" : "")
-		);
-		expect(await isGitConfigured()).toBe(false);
-	});
-
-	test("no email", async () => {
-		const spy = vi.spyOn(command, "runCommand");
-		spy.mockImplementation((cmd) =>
-			promisify(cmd.includes("name") ? "test user" : "")
-		);
-		expect(await isGitConfigured()).toBe(false);
-	});
-
-	test("runCommand fails", async () => {
-		const spy = vi.spyOn(command, "runCommand");
-		spy.mockImplementation(() => {
-			throw new Error("git not found");
-		});
-		expect(await isGitConfigured()).toBe(false);
-	});
 });
 
 describe("isUpdateAvailable", () => {
